LayoutTests/svg/dom/SVGViewSpec-invalid-ref-crash-expected.txt

Confirm that exceptions are thrown, or empty string returned, when an SVGViewSpec is used after its corresponding element has been removed.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS viewPreserveAspectRatio = svgView.preserveAspectRatio; threw exception Error: InvalidStateError: DOM Exception 11.
PASS viewTransform = svgView.transform; is null
PASS viewViewTarget = svgView.viewTarget; threw exception Error: InvalidStateError: DOM Exception 11.
PASS viewViewBox = svgView.viewBox; threw exception Error: InvalidStateError: DOM Exception 11.
PASS successfullyParsed is true

TEST COMPLETE
